refactor: `mora_list`を`mora_mappings`に改名
