+++ crates/voicevox_core/src/engine.rs
@@ -2,7 +2,7 @@
 
 mod acoustic_feature_extractor;
 mod audio_file;
-mod mora_list;
+mod mora_mappings;
 mod sampling_rate;
 pub(crate) mod talk;
